+++ src/classic/pbap_client.h
@@ -76,6 +76,13 @@ uint8_t pbap_disconnect(uint16_t pbap_cid);
 uint8_t pbap_set_phonebook(uint16_t pbap_cid, const char * path);
 
 /**
+ * @brief Get size of phone book from PSE
+ * @param pbap_cid
+ * @return status
+ */
+uint8_t pbap_get_phonebook_size(uint16_t pbap_cid);
+
+/**
  * @brief Pull phone book from PSE
  * @param pbap_cid
  * @return status